STELLOPT: Allow _opt files for automated zipping in some scenarios.

+++ STELLOPTV2/Sources/General/stellopt_optimize.f90
@@ -23,7 +23,7 @@
 !----------------------------------------------------------------------
       IMPLICIT NONE
       !LOGICAL ::  lrestart
-      LOGICAL ::  lfile_exists, lskip_min
+      LOGICAL ::  lfile_exists, lskip_min, ldeleteopt
       INTEGER ::  ier, iunit,nvar_in, nprint, info, ldfjac,nfev,&
                   iunit_restart, nfev_save, npop, ndiv, i
       INTEGER, ALLOCATABLE :: ipvt(:)
@@ -145,6 +145,7 @@
 
       ! DEFAULT
       lskip_min = .false.
+      ldeleteopt = .TRUE.
 
       ! Do runs
       SELECT CASE(TRIM(opt_type))
@@ -181,10 +182,12 @@
                        qtf, wa1, wa2, wa3, wa4,vars_min,vars_max)
             DEALLOCATE(ipvt, qtf, wa1, wa2, wa3, wa4, fjac)
          CASE('eval_xvec')
+            ldeleteopt = .FALSE.
             lskip_min = .true.
             CALL xvec_eval(stellopt_fcn,nvars,mtargets,xvec_file)
          CASE('one_iter','single','eval','single_iter')
             lskip_min = .true.
+            ldeleteopt = .FALSE.
             ALLOCATE(fvec(mtargets))
             fvec     = 0.0
             info     = FLAG_SINGLETASK
@@ -201,6 +204,7 @@
             IF (myid == master) info = flag_cleanup_lev
             call stellopt_fcn(mtargets, nvars, vars, fvec, info, nfev)
          CASE('one_iter_norm')
+            ldeleteopt = .FALSE.
             ALLOCATE(fvec(mtargets))
             fvec     = 0.0
             info     = FLAG_SINGLETASK
@@ -235,6 +239,7 @@
             CLOSE(iunit)
             CLOSE(iunit_restart)
          CASE('map')
+            ldeleteopt = .FALSE.
             lskip_min = .true.
             nprint = 6
             lno_restart = .true.
@@ -245,6 +250,7 @@
                 WRITE(6,*) '       See map.dat for data               '
             END IF
          CASE('map_linear')
+            ldeleteopt = .FALSE.
             lskip_min = .true.
             nprint = 6
             lno_restart = .true.
@@ -255,6 +261,7 @@
                 WRITE(6,*) '       See map.dat for data               '
             END IF
          CASE('map_plane')
+            ldeleteopt = .FALSE.
             lskip_min = .true.
             nprint = 6
             lno_restart = .true.
@@ -265,6 +272,7 @@
                 WRITE(6,*) '       See map.dat for data               '
             END IF
          CASE('map_hypers')
+            ldeleteopt = .FALSE.
             lskip_min = .true.
             ALLOCATE(wa1(nvars),fvec(mtargets))
             wa1 = vars
@@ -316,11 +324,13 @@
       END IF
 !DEC$ IF DEFINED (STELZIP)
       IF (myid == master) THEN
-         ! Remove the *_opt* files
-         WRITE(6,*) ' Cleaning up _opt files'; CALL FLUSH(6); ier = 0; ierr_mpi = 0; cmdtxt=''
-         CALL EXECUTE_COMMAND_LINE("rm -rf *_opt*",WAIT=.TRUE.,EXITSTAT=ier,CMDSTAT=ierr_mpi,CMDMSG=cmdtxt)
-         WRITE(6,*) ' rm: EXITSTAT=',ier,' CMDSTAT=',ierr_mpi; CALL FLUSH(6)
-         WRITE(6,*) '     MESSAGE: ',TRIM(cmdtxt); CALL FLUSH(6)
+         IF (ldeleteopt) THEN
+            ! Remove the *_opt* files
+            WRITE(6,*) ' Cleaning up _opt files'; CALL FLUSH(6); ier = 0; ierr_mpi = 0; cmdtxt=''
+            CALL EXECUTE_COMMAND_LINE("rm -rf *_opt*",WAIT=.TRUE.,EXITSTAT=ier,CMDSTAT=ierr_mpi,CMDMSG=cmdtxt)
+            WRITE(6,*) ' rm: EXITSTAT=',ier,' CMDSTAT=',ierr_mpi; CALL FLUSH(6)
+            WRITE(6,*) '     MESSAGE: ',TRIM(cmdtxt); CALL FLUSH(6)
+         END IF
          ! Zip up the results
          WRITE(6,*) ' Zipping files'; CALL FLUSH(6); ier = 0; ierr_mpi = 0; cmdtxt=''
          CALL EXECUTE_COMMAND_LINE("zip -r stellopt_files.zip *",WAIT=.TRUE.,EXITSTAT=ier,CMDSTAT=ierr_mpi,CMDMSG=cmdtxt)
